factorize serializer and remove JSONOutputter from ConsoleOutputter ancestors
