+++ lib/inferno/apps/cli/execute/json_outputter.rb
@@ -1,12 +1,13 @@
-require 'active_support'
-require_relative '../../web/serializers/test_run'
-require_relative '../../web/serializers/result'
+require_relative 'serialize'
 
 module Inferno
   module CLI
     class Execute
       # @private
       class JSONOutputter
+
+        include Serialize
+
         def print_start_message(_options); end
 
         def print_around_run(_options, &)
@@ -21,15 +22,6 @@ module Inferno
 
         def print_error(_options, exception)
           puts exception.to_json
-        end
-
-        def serialize(entity)
-          case entity.class.to_s
-          when 'Array'
-            JSON.pretty_generate(entity.map { |item| JSON.parse serialize(item) })
-          else
-            Inferno::Web::Serializers.const_get(entity.class.to_s.demodulize).render(entity)
-          end
         end
       end
     end
